Serialize DateTime
